Always check if it's time to trim an EMP hash
The previous logic was to only trim the hash table if a new hash entry
was being created.  Usually this is a valid assumption but, in the case
of lphn, it's plausible that only a single user posts to a group.  In
this case, a single hash entry will grown untrimmed until messages are
rejected.  This change ensures a trim check takes place regardless of if
the table is being extended or not.  My thanks to Julien Élie for
catching it.