Fix the detection of empty header fields
The art dictionary contains all the usual header fields (more than 70,
populated by innd).  When empty, they have a None value.

In some places in the code, the detection was wrong (looking at the presence
of the header field name in the dictionary instead of the presence of a value).